+++ packages/bridging/src/providers/across/util.ts
@@ -5,8 +5,8 @@ import { AcrossQuoteResult } from './AcrossBridgeProvider'
 import { ACROSS_DEPOSIT_EVENT_INTERFACE, COW_TRADE_EVENT_INTERFACE } from './const/interfaces'
 import { ACROSS_TOKEN_MAPPING, AcrossChainConfig } from './const/tokens'
 import { ACROSS_SPOOK_CONTRACT_ADDRESSES } from './const/contracts'
-import { COW_PROTOCOL_SETTLEMENT_CONTRACT_ADDRESS, SupportedChainId, TargetChainId, isEvmChain } from '@cowprotocol/sdk-config'
-import { getBigNumber, OrderKind } from '@cowprotocol/sdk-order-book'
+import { COW_PROTOCOL_SETTLEMENT_CONTRACT_ADDRESS, SupportedChainId, TargetChainId } from '@cowprotocol/sdk-config'
+import { OrderKind } from '@cowprotocol/sdk-order-book'
 import { getGlobalAdapter, Log } from '@cowprotocol/sdk-common'
 import stringify from 'json-stable-stringify'
 
@@ -30,9 +30,6 @@ export function getChainConfigs(
 }
 
 function getChainConfig(chainId: TargetChainId): AcrossChainConfig | undefined {
-  if (!isEvmChain(chainId)) {
-    return undefined
-  }
   return ACROSS_TOKEN_MAPPING[chainId]
 }
 
@@ -77,9 +74,9 @@ function toAmountsAndCosts(
   const { amount, sellTokenDecimals, buyTokenDecimals } = request
 
   // Get the amounts before fees
-  const sellAmountBeforeFeeBig = getBigNumber(amount, sellTokenDecimals)
-  const sellAmountBeforeFee = sellAmountBeforeFeeBig.big
-  const buyAmountBeforeFee = getBigNumber(sellAmountBeforeFeeBig.num, buyTokenDecimals).big // Sell and buy token should be the same asset for current implementation, but technically they can have different decimals
+  const sellAmountBeforeFee = BigInt(amount)
+  // Sell and buy token should be the same asset for current implementation, but technically they can have different decimals
+  const buyAmountBeforeFee = (sellAmountBeforeFee * 10n ** BigInt(buyTokenDecimals)) / 10n ** BigInt(sellTokenDecimals)
 
   // Apply the fee to the buy amount (sell amount doesn't change)
   const totalRelayerFeePct = BigInt(suggestedFees.totalRelayFee.pct)