+++ packages/bridging/src/providers/bungee/const/contracts.ts
@@ -16,6 +16,7 @@ export const BungeeCowswapLibAddresses: Record<TargetChainId, string | undefined
   [SupportedChainId.SEPOLIA]: undefined,
   [AdditionalTargetChainId.OPTIMISM]: BUNGEE_COWSWAP_LIB_ADDRESS,
   [AdditionalTargetChainId.BITCOIN]: undefined,
+  [AdditionalTargetChainId.SOLANA]: undefined,
 }
 
 const SOCKET_VERIFIER_ADDRESS = '0xa27A3f5A96DF7D8Be26EE2790999860C00eb688D'
@@ -34,6 +35,7 @@ export const SocketVerifierAddresses: Record<TargetChainId, string | undefined> 
   [SupportedChainId.SEPOLIA]: undefined,
   [AdditionalTargetChainId.OPTIMISM]: SOCKET_VERIFIER_ADDRESS,
   [AdditionalTargetChainId.BITCOIN]: undefined,
+  [AdditionalTargetChainId.SOLANA]: undefined,
 }
 
 const BUNGEE_APPROVE_AND_BRIDGE_V1_ADDRESS = '0xD06a673fe1fa27B1b9E5BA0be980AB15Dbce85cc'
@@ -52,4 +54,5 @@ export const BungeeApproveAndBridgeV1Addresses: Record<TargetChainId, string | u
   [SupportedChainId.SEPOLIA]: undefined,
   [AdditionalTargetChainId.OPTIMISM]: BUNGEE_APPROVE_AND_BRIDGE_V1_ADDRESS,
   [AdditionalTargetChainId.BITCOIN]: undefined,
+  [AdditionalTargetChainId.SOLANA]: undefined,
 }