+++ packages/common/src/utils/token.ts
@@ -1,6 +1,6 @@
 import { Nullish } from '../types'
-import { ChainId, getChainInfo, isEvmChain, isNonEvmChain, isSupportedChain, WRAPPED_NATIVE_CURRENCIES } from '@cowprotocol/sdk-config'
-import { AddressKey, getAddressKey, getEvmAddressKey, isBtcAddress, isEvmAddress } from './address'
+import { ChainId, getChainInfo, isSupportedChain, WRAPPED_NATIVE_CURRENCIES } from '@cowprotocol/sdk-config'
+import { AddressKey, getAddressKey, getEvmAddressKey, isEvmAddress } from './address'
 
 interface TokenLike {
   chainId: ChainId
@@ -35,40 +35,17 @@ export function areAddressesEqual(a: Nullish<string>, b: Nullish<string>): boole
     return getEvmAddressKey(a) === getEvmAddressKey(b)
   }
 
-  const aIsBtc = isBtcAddress(a)
-  const bIsBtc = isBtcAddress(b)
-
-  if (aIsBtc && bIsBtc) {
-    return a === b
-  }
-
-  return false
+  // sol and btc addresses are already in the correct format
+  return a === b
 }
 
 export function isNativeToken(token: TokenLike): boolean {
   const chainInfo = getChainInfo(token.chainId)
-  if (!chainInfo) {
-    return false
-  }
-
-  // For EVM chains, check if the token address matches the native currency address
-  if (isEvmChain(chainInfo.id)) {
-    if ('address' in chainInfo.nativeCurrency) {
-      return areAddressesEqual(chainInfo.nativeCurrency.address, token.address)
-    }
-    return false
-  }
-
-  // For non-EVM chains, if there's no address (empty string), it's the native token
-  if (isNonEvmChain(chainInfo.id)) {
-    return token.address === '' // todo need to adjust token definition type
-  }
-
-  return false
+  return areAddressesEqual(chainInfo?.nativeCurrency.address, token.address)
 }
 
 export function isWrappedNativeToken(token: TokenLike): boolean {
-  if (!isEvmChain(token.chainId) || !isSupportedChain(token.chainId)) {
+  if (!isSupportedChain(token.chainId)) {
     return false
   }
   return areAddressesEqual(WRAPPED_NATIVE_CURRENCIES[token.chainId]?.address, token.address)