+++ packages/config/src/constants/addresses.ts
@@ -1,4 +1,14 @@
 /**
  * The standard address used to represent native currency (ETH, MATIC, etc.) in EVM chains
  */
-export const NATIVE_CURRENCY_ADDRESS = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE'
+export const EVM_NATIVE_CURRENCY_ADDRESS = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE'
+
+/**
+ * There is no standard address for SOL, so we use a default program address
+ */
+export const SOL_NATIVE_CURRENCY_ADDRESS = '11111111111111111111111111111111'
+
+/**
+ * There is no standard address for BTC, so we use the Genesis address as a token address
+ */
+export const BTC_CURRENCY_ADDRESS = '1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa'